backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.10.0
aiofiles>=24.1.0
SQLAlchemy>=2.0.23
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
rapidfuzz>=3.5.0
